Create Documentation for motion matching by drew
This documentation was originally written five years ago for Motion Matching 1.0, so some parts may be outdated or inconsistent. Still, it can be helpful for understanding the system. Please be patient with any such issues.

This translation is from Polish to English, so some language gaps may occur.
All the best!






 Caution: I am not the author of this system. The PDF originally existed, but for some reason, the author removed it.

In the YouTube video about this system, many people were struggling to understand it and were asking for documentation.

After a bit of searching, I found this PDF online (from version 1.0). So, I'm uploading it here to help everyone out.
Hope it helps!